Move IndexMetadata, NodesResponse, SearchShardsResponse, Shard, ElasticsearchNode into Elasticsearch client directory
