+++ presto-elasticsearch/src/main/java/com/facebook/presto/elasticsearch/ElasticsearchClient.java
@@ -17,6 +17,11 @@ import com.facebook.airlift.json.JsonCodec;
 import com.facebook.airlift.json.ObjectMapperProvider;
 import com.facebook.airlift.log.Logger;
 import com.facebook.airlift.security.pem.PemReader;
+import com.facebook.presto.elasticsearch.client.ElasticsearchNode;
+import com.facebook.presto.elasticsearch.client.IndexMetadata;
+import com.facebook.presto.elasticsearch.client.NodesResponse;
+import com.facebook.presto.elasticsearch.client.SearchShardsResponse;
+import com.facebook.presto.elasticsearch.client.Shard;
 import com.facebook.presto.spi.PrestoException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -99,7 +104,7 @@ public class ElasticsearchClient
     private final int scrollSize;
     private final Duration scrollTimeout;
 
-    private final AtomicReference<Set<Node>> nodes = new AtomicReference<>(ImmutableSet.of());
+    private final AtomicReference<Set<ElasticsearchNode>> nodes = new AtomicReference<>(ImmutableSet.of());
     private final ScheduledExecutorService executor = newSingleThreadScheduledExecutor(daemonThreadsNamed("NodeRefresher"));
     private final AtomicBoolean started = new AtomicBoolean();
     private final Duration refreshInterval;
@@ -138,10 +143,10 @@ public class ElasticsearchClient
     {
         // discover other nodes in the cluster and add them to the client
         try {
-            Set<Node> nodes = fetchNodes();
+            Set<ElasticsearchNode> nodes = fetchNodes();
 
             HttpHost[] hosts = nodes.stream()
-                    .map(Node::getAddress)
+                    .map(ElasticsearchNode::getAddress)
                     .map(address -> HttpHost.create(format("%s://%s", tlsEnabled ? "https" : "http", address)))
                     .toArray(HttpHost[]::new);
 
@@ -293,36 +298,36 @@ public class ElasticsearchClient
         }
     }
 
-    private Set<Node> fetchNodes()
+    private Set<ElasticsearchNode> fetchNodes()
     {
         NodesResponse nodesResponse = doRequest("_nodes/http", NODES_RESPONSE_CODEC::fromJson);
 
-        ImmutableSet.Builder<Node> result = ImmutableSet.builder();
+        ImmutableSet.Builder<ElasticsearchNode> result = ImmutableSet.builder();
         for (Map.Entry<String, NodesResponse.Node> entry : nodesResponse.getNodes().entrySet()) {
             String nodeId = entry.getKey();
             NodesResponse.Node node = entry.getValue();
 
             if (node.getRoles().contains("data")) {
-                result.add(new Node(nodeId, node.getHttp().getAddress()));
+                result.add(new ElasticsearchNode(nodeId, node.getHttp().getAddress()));
             }
         }
         return result.build();
     }
 
-    public Set<Node> getNodes()
+    public Set<ElasticsearchNode> getNodes()
     {
         return nodes.get();
     }
 
     public List<Shard> getSearchShards(String index)
     {
-        Map<String, Node> nodeById = getNodes().stream()
-                .collect(toImmutableMap(Node::getId, Function.identity()));
+        Map<String, ElasticsearchNode> nodeById = getNodes().stream()
+                .collect(toImmutableMap(ElasticsearchNode::getId, Function.identity()));
 
         SearchShardsResponse shardsResponse = doRequest(format("%s/_search_shards", index), SEARCH_SHARDS_RESPONSE_CODEC::fromJson);
 
         ImmutableList.Builder<Shard> shards = ImmutableList.builder();
-        List<Node> nodes = ImmutableList.copyOf(nodeById.values());
+        List<ElasticsearchNode> nodes = ImmutableList.copyOf(nodeById.values());
 
         for (List<SearchShardsResponse.Shard> shardGroup : shardsResponse.getShardGroups()) {
             Stream<SearchShardsResponse.Shard> preferred = shardGroup.stream()
@@ -333,7 +338,7 @@ public class ElasticsearchClient
                     .findFirst();
 
             SearchShardsResponse.Shard chosen;
-            Node node;
+            ElasticsearchNode node;
             if (candidate.isPresent()) {
                 chosen = candidate.get();
                 node = nodeById.get(chosen.getNode());
