+++ presto-elasticsearch/src/main/java/com/facebook/presto/elasticsearch/client/ElasticsearchNode.java
@@ -11,16 +11,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.elasticsearch;
+package com.facebook.presto.elasticsearch.client;
 
 import static java.util.Objects.requireNonNull;
 
-public class Node
+public class ElasticsearchNode
 {
     private final String id;
     private final String address;
 
-    public Node(String id, String address)
+    public ElasticsearchNode(String id, String address)
     {
         this.id = requireNonNull(id, "id is null");
         this.address = requireNonNull(address, "address is null");
