+++ presto-elasticsearch/src/main/java/com/facebook/presto/elasticsearch/client/SearchShardsResponse.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.elasticsearch;
+package com.facebook.presto.elasticsearch.client;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
